* master: (1397 commits)
  svn update
  update addon scripts
  patch [#28218] During-render callback functionality from Jesse Kaukonen (gekko)
  COLLADA: Take parent bone length and direction instead of using bone pointing up with length 1. Looks much nicer and less confusing on larger armatures now.
  BGE: Adding two new functions to bge.render to allow users to change the anisotropic filtering level used by textures:   * setAnisotropicFiltering(level)   * getAnisotropicFiltering()
  typo fix: end of lines ;; --> ;
  fix [#28430] Image with Stampinfo does not get saved correctly with alpha
  OSX: additionally choice to link against python-framework again
  fix for error in my recent change to image save. - relative path wasn't being made absolute. - saving renders was always defaulting to multilayer exr, now use the output format set.
  catch exception and report an error when failing to write exr files - was crashing with debug builds.
  SVN maintenance.
  When creating a multisample window for Win32, the first context created, which checks for multisample support, was never being deleted.
  Check for potential crasher.
  Fix #28427: smooth faces flash momentarily when extruded using "extrude and move on normals" tool (E key)
  Fix for [#28425] when user prefs -> editing -> align to == "view" newly inserted objects do not show the applied rotation in the tools panel
  2.5 Game UI: *Fix for clutter after pepper merge, 3 booleans in one row, is 1 too much here ;-)
  Enable libsndfile by default on win32 too.
  Fix for my last commit.
  minor edits, pep8 - also correct float -> double promotion for blf.
  2.5 UI Files: * Code cleanup after Pepper merge.
  ...

Conflicts:
	source/blender/blenkernel/CMakeLists.txt
	source/blender/makesrna/intern/rna_modifier.c
	source/blender/makesrna/intern/rna_texture.c
	source/blender/modifiers/CMakeLists.txt

+++ source/blender/editors/include/ED_gpencil.h
@@ -1,6 +1,4 @@
 /*
- * $Id$
- *
  * ***** BEGIN GPL LICENSE BLOCK *****
  *
  * This program is free software; you can redistribute it and/or
@@ -56,7 +54,7 @@ struct wmKeyConfig;
  * Used as part of the 'stroke cache' used during drawing of new strokes
  */
 typedef struct tGPspoint {
-	short x, y;				/* x and y coordinates of cursor (in relative to area) */
+	int x, y;				/* x and y coordinates of cursor (in relative to area) */
 	float pressure;			/* pressure of tablet at this point */
 } tGPspoint;
